feat: refactor composio connectors for modularity
